+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Features
 
-- Migrate to datapack for blacklist.
+- Support 1.18 by using a `java.util.Random` in place of a `RandomGenerator`.
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/TotallyBalancedBoneDrops/compare/1.0.0...1.1.0)
+[Full Changelog](https://github.com/JamCoreModding/TotallyBalancedBoneDrops/compare/1.1.0...1.1.1)